API/testgenai-master/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
annoy==1.17.3
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitarray==2.9.2
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cohere==4.52
colbert-ai==0.2.19
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
datasets==2.18.0
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fast-pytorch-kmeans==0.2.0.1
fastapi==0.109.0
fastavro==1.9.4
filelock==3.13.1
fiona==1.9.5
Flask==3.0.3
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.12.2
geopandas==0.14.3
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.16.1
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
groq==0.4.2
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.5
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langchain-google-genai==0.0.6
langchain-openai==0.0.6
langsmith==0.0.85
lida==0.0.14
llama-index==0.10.27
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.11
llama-index-core==0.10.27
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.14
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.16
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.0
llamaindex-py-client==0.1.16
llmx==0.0.21a0
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-venn==0.11.10
mizani==0.11.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.3
onnx==1.16.0
openai==1.16.2
packaging==23.2
pandas==2.2.0
patsy==0.5.6
pillow==10.2.0
plotly==5.19.0
plotnine==0.13.0
proto-plus==1.23.0
protobuf==4.25.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.0
pydantic_core==2.16.1
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pynvml==11.5.0
pyodbc==5.0.1
pyparsing==3.1.1
pypdf==4.2.0
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.4
PyYAML==6.0.1
RAGatouille==0.0.8.post2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==2.3.1
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
srsly==2.4.8
sseclient-py==1.8.0
starlette==0.35.1
statsmodels==0.14.1
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.6.0
together==0.2.11
tokenizers==0.15.1
toolz==0.12.1
torch==2.2.0
tqdm==4.66.1
transformers==4.37.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
voyager==2.0.6
Werkzeug==3.0.2
wordcloud==1.9.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
